deployment_scripts/orchestrator_requirements.txt
boto3>=1.34.0
botocore>=1.34.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0 